+++ src/components/TodoListItem.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import style from "./TodoListItem.module.css";
+import PropTypes from 'prop-types';
 
 
 const TodoListItem = ({ todo, onRemoveTodo }) => {
@@ -12,6 +13,12 @@ const TodoListItem = ({ todo, onRemoveTodo }) => {
             </li>
         </>
     );
+};
+
+TodoListItem.propTypes = {
+    id: PropTypes.string,
+    onRemoveTodo: PropTypes.func,
+    title: PropTypes.object,
 };
 
 export default TodoListItem;
